Some Module 5 fixes
