+++ docs/aca/05-aca-dapr-pubsubapi/TasksNotifierController.cs
@@ -1,7 +1,5 @@
 using Dapr.Client;
 using Microsoft.AspNetCore.Mvc;
-using SendGrid;
-using SendGrid.Helpers.Mail;
 using TasksTracker.Processor.Backend.Svc.Models;
 
 namespace TasksTracker.Processor.Backend.Svc.Controllers
@@ -13,7 +11,7 @@ namespace TasksTracker.Processor.Backend.Svc.Controllers
         private readonly IConfiguration _config;
         private readonly ILogger _logger;
         private readonly DaprClient _daprClient;
-        
+
         public TasksNotifierController(IConfiguration config, ILogger<TasksNotifierController> logger, DaprClient daprClient)
         {
             _config = config;
@@ -23,38 +21,12 @@ namespace TasksTracker.Processor.Backend.Svc.Controllers
 
         [Dapr.Topic("dapr-pubsub-servicebus", "tasksavedtopic")]
         [HttpPost("tasksaved")]
-        public async Task<IActionResult> TaskSaved([FromBody] TaskModel taskModel)
+        public Task<IActionResult> TaskSaved([FromBody] TaskModel taskModel)
         {
+            var msg = string.Format("Started processing message with Task Name '{0}'", taskModel.TaskName);
             _logger.LogInformation("Started processing message with Task Name '{0}'", taskModel.TaskName);
 
-            var sendGridResponse = await SendEmail(taskModel);
-
-            if (sendGridResponse.Item1)
-            {
-                return Ok($"SendGrid response status code: {sendGridResponse.Item1}");
-            }
-
-            return BadRequest($"Failed to send email, SendGrid response status code: {sendGridResponse.Item1}");
-        }
-
-        private async Task<Tuple<bool, string>> SendEmail(TaskModel taskModel)
-        {
-            var apiKey = _config.GetValue<string>("SendGrid:ApiKey");
-            var sendEmailResponse = true;
-            var sendEmailStatusCode = System.Net.HttpStatusCode.Accepted;
-            var client = new SendGridClient(apiKey);
-            var from = new EmailAddress("taiseer.joudeh@gmail.com", "Tasks Tracker Notification");
-            var subject = $"Task '{taskModel.TaskName}' is assigned to you!";
-            var to = new EmailAddress(taskModel.TaskAssignedTo, taskModel.TaskAssignedTo);
-            var plainTextContent = $"Task '{taskModel.TaskName}' is assigned to you. Task should be completed by the end of: {taskModel.TaskDueDate.ToString("dd/MM/yyyy")}";
-            var htmlContent = plainTextContent;
-            var msg = MailHelper.CreateSingleEmail(from, to, subject, plainTextContent, htmlContent);
-
-            var response = await client.SendEmailAsync(msg);
-            sendEmailResponse = response.IsSuccessStatusCode;
-            sendEmailStatusCode = response.StatusCode;
-
-            return new Tuple<bool, string>(sendEmailResponse, sendEmailStatusCode.ToString());
+            return OK(msg);
         }
     }
 }
