+++ docs/aca/13-appendix/03-variables.md
@@ -13,6 +13,7 @@ AZURE_SUBSCRIPTION_ID               : $AZURE_SUBSCRIPTION_ID
 BACKEND_API_EXTERNAL_BASE_URL       : $BACKEND_API_EXTERNAL_BASE_URL
 BACKEND_API_INTERNAL_BASE_URL       : $BACKEND_API_INTERNAL_BASE_URL
 BACKEND_API_NAME                    : $BACKEND_API_NAME
+BACKEND_SERVICE_APP_PORT            : $BACKEND_SERVICE_APP_PORT
 BACKEND_SERVICE_NAME                : $BACKEND_SERVICE_NAME
 BACKEND_SERVICE_PRINCIPAL_ID        : $BACKEND_SERVICE_PRINCIPAL_ID
 COSMOS_DB_ACCOUNT                   : $COSMOS_DB_ACCOUNT
